requirements.txt
streamlit==1.32.0
scikit-learn==1.4.0
xgboost==2.0.0
lightgbm==4.0.0
shap==0.41.0
pandas==2.2.0
numpy==1.24.0
matplotlib==3.7.0
seaborn==0.13.0
PyYAML==6.0
fpdf2==2.7.4
optuna==3.6.0
scipy==1.11.0
imblearn